added a Cashu QR Code
I created this QR Code of Cashu and added it to a sign that says how to redeem it. The cashu can only be redeemed once, but I am using it as a template. Each card will need a new QR code unless we can get a Cashu fountain of some sort.